Jfreechart + ScreenSize + metro.txt
Jfreechart is completely removed from the project and only imported from
maven pom.xml
The screensize is automatically adapted from one to another, no need of
resizing nor repositioning the window.
The metro.txt is now pre-calculated and so faster to reload.

I change the name of the contributor as I change the computer...
Signed-off-by: Pery Dorian <dorian.pery@gmail.com>